edit submitted events
